pyproject.toml
[project]
name = "credit-risk-scoring"
version = "0.1.0"
description = "Credit risk scoring pipeline with Sklearn and PyTorch"
author = [
    {name = "Javier Alberto Juarez Luna"}
]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "torch>=2.2",
    "ipykernel>=7.2.0",
    "pandera>=0.20",
    "mlflow>=2.12",
    "streamlit>=1.35",
    "dvc>=3.50",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "pre-commit>=3.7",
]


[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.per-file-ignores]
"notebooks/*.ipynb" = ["B018","E402","I001","E501"]
